Move duplicated functions into utils
